added configurable custom HTTP header support, it is now possible to add or remove custom headers to all HTTP responses (this also includes ssdp/gena messages)
the funtions operate similar to the virtual directory 
implementation. as I see it - this is not threadsafe.

those functions (also the already existing virtual directory
functoins like add/remove) should never be called during
operation, epsceially never call the remove functions after
you have registered your root device

the headers are not yet added to the client HTTP messages.

upnpapi.c - fixed a bug where virtual directory could not
be removed if it was first in the list

found some strcpy's instead of strncpy, actually quite
a lot of dangerous places; will have to go through the
existing code and fix that.