we are now handling the rpl_realloc case
AC_FUNC_MALLOC and AC_FUNC_REALLOC will redefine malloc and realloc if
they think that we are not building for the GNU C library. This can be
a problem for cross compile environments, so our configure script now has
an option to disable this feature.

However, we also added appropriate rpl_ functions to handle the cases
where malloc is indeed not GNU C compatible.