[PATCH] Add the cache feature of ffmpegthumbnailer.
I want the cache feature of ffmpegthumbnailer because my machine
is not powerful :-(
So I created this patch for the cache feature.

This patch adds a new option "cache-dir" in config.xml.
If not specifying any string, the cache feature is disable.
And specifying some directory, the cache feature is enable
and the cache files will be created under the directory.

Signed-off-by: Ken'ichi Ohmichi <ken1ohmichi@gmail.com>