Fix playlists losing their order
Playlists lose their order, with the files just shown in alphabetical
order. This is despite the correct ordinals going into the database.
The playlist simply wasn't being marked as a playlist, so these were
ignored.

There is an extra "containerclass = ts;". If this line is removed,
playlists work correctly. It seems to have been left in from before the
conditional statement preceding it was added.