Fix anoying logging of transcoder arguments
Mediatomb logs the transcoders commands arguments as Info when it really
should be Debug. Also the location and command used is not logged which
makes the logging quite "unspecific". The attached patch changes the
arguments log to debug and adds debug logs for command used and the
location of the source file.