Fix timed autoscans
Added a addTimerSubscriber() call for each item list of dirs.
Moved some of the timer code from the header to cpp file and reorganised the
timer code a bit.

Moved TimerParameter to inside timer and used it instead of object. SQLite was
using Ref<Object> so swapped it to impement timerSubscriber directly.

Other cleanups including remove compareTimespecs(), rename notified() and
whitespace and formatting.