Give Timer its own thread.
Timer has since been refactored and this introduced breakage in the main class
due the interesting coupling in vanilla where timer's triggerWait() would be
called repeatedly by the main loop.

This change gives Timer its own thread. This gives cleaner separation between
it and the main signal handling logic. Main now has a simple condition which it
can wait on until it gets a signal.

This fixes #53. Thanks to @whorfin for the ongoing testing and issue reporting!