Removing sudo dependecy and fixing shellcheck warnings
Remove the sudo dependecy in the install scripts and added a
if-statement to check if the script is run with superuser access to
    perform the install.
This removes the need to have sudo installed and also has a better
error-handeling if the user types in the root password wrong 3 times.

Fixed shellcheck warnings which should result in robuster code.