Bugfixing handling OBJECT_TYPE_EXTERNAL_URL
Validating Objects: now ExtUrls are no longer rejected (they were rejected because no file named like the url was found)
Scripting: just like above
UPNP_XML: if you use transcoding with an EXTERNAL_URL-Type and hide the original resource the server's ip and port is now added to the <res>-Tag correctly.
With this you are able to use gerbera like the old mediatomb before to import and play cue-files. A how-to is added to the docs. IMHO this is a little dirty hack. The best way is to add a new object type for such a kind of use the quite powerful transcoding facility.