fix broken for loop breaking artist display
0242666a65befc51832dd057f2d2b9e22469882c broke this by incorrectly
converting several statements to a loop. What happens with that commit
is that the second vector member does not even get looked at.

Change so that it's a simple loop. Even though two members are the same,
it seems to work properly as before.

Also cleaned up the surrounding code.

Signed-off-by: Rosen Penev <rosenp@gmail.com>