Concatenate multivalue field with entrySeparator
Fix: Concatenate multivalue field with entrySeparator and not space

In ID3v2.4 tags multiple values are allowed (see also #1425).
These values are stored as null separated list in the mp3 file tag.
If taglib is called with textFrame->toString() the multiple values are concatenated with a space.
To concatenate these values with the defined entrySeparater this call is changed to textFrame->fieldList().toString(entrySeparator).

For TXXX tags the first element in the fieldList is the subtag name and extracted separately. The other entries are concatenated with the entrySeparator.