Add M3U8 support
The M3U8 format allows special characters within the file names and
descriptions. It uses a UTF8 BOM to indicate the byte order. This commit
strips the BOM in common.js to make sure the rest of the file is
interpreted correctly.
It also adds the necessary tests to verify that the BOM is correctly
stripped and that the special characters are stored as UTF-8.