Remove duplicate logic from file_request_handler
We were doing a lot of stuff twice due to being unable to keep data
between the the getInfo() and open() calls.

Now that our handler is used accross both we can do all the work just
once in getInfo() and make the IOHandler available for open() which now
contains basically zero logic.

Still lots of room for improvements, but much better than it was.