Refactor file_request_handler
Reverts the changes to cache stuff across get_info()/open() calls, this
is not feasable because libupnp may not call open() in the case of HEAD
requests, giving us no opportunity to clean up resources.

This sucks because we have to parse/load everything twice per request,
but it should stop transcoding processes being leaked.