Drop resourceHandler concept
Nowadays thanks to Karls great work we no longer have the virtual
resource mess to deal with as things like album art now live in
the DB, so we can simplify this code a bit.

Fixes broken artwork.

In fact with this change the `rct` handler is also ignored, just
`res_id/1` is enough to get the image.