Expand firewall advice in docs (#2993)
Add a new Firewall section to the Troubleshooting page.

Expand the Port section of the Troubleshooting page to complement the
new Firewall section.

Add info about Fedora Server edition on the Installing Gerbera page.
The existing Fedora section was ok for the Workstation edition. The
added info should save a user a bunch of troubleshooting if they
install on the Server edition.