matroska_handler: update API calls to work with libebml/libmatroska 2.0 (#3007)
* matroska_handler: fix include with libebml/libmatroska 2.0

* matroska_handler: use libebml/libmatroska namespaces

They used to be forced when using the headers but no longer in 2.0 libraries.

* matroska_handler: fix namespaces usage

In 2.0 the namespace defines are gone, as the namespace won't change.

The using namespace libebml is not done in headers anymore.

* matroska_handler: fix return types for IOCallback derived class

Standard C types are used. The old custom ones were the same anyway.

* matroska_handler: fix custom types usage

In v1 and v2 of libebml they are equivalent to the standard C types.
In v2 the custom types are gone.

* matroska_handler: use EBML_ID() to get the EbmlId of a class

It works with v1 and v2 versions of libebml.

* matroska_handler: use EBML_CLASS_CONTEXT() to get the context of KaxSegment

It works with v1 and v2 versions of libebml.

* matroska_handler: use EBML_INFO() to get the EbmlCallbacks of KaxSegment

It works with v1 and v2 versions of libebml.

* matroska_handler: use EbmlMaster::GetDataStart to get the data position

It works with v1 and v2 versions of libebml.

* matroska_handler: use KaxSeek::IsEbmlId to check a seek is a given type

It works with v1 and v2 versions of libebml.

* matroska_handler: use EbmlId::FromBuffer to construct an EbmlId in libebml v2