Server: Handle all HTTP via virtualdir
Previously we let libupnp serve $WEBROOT directly, this was used for
the web ui and the various XML definitions.

This change moves to serving those files via virtualdir instead. This
gives a lot more control. e.g. We can stop serving the UI if is disabled
in config, instead of implmenting that in the UI JS.

It also means we can serve (in the future) Qurik'd versions of our XML
deinitions, which is a pretty interesting idea.