Do not use std::prev with fs::path
This fixed crashes on freebsd 14.1
I was able to trace the cause back to this bug reports in libc++: https://github.com/llvm/llvm-project/issues/55521
As far as I understand, they are not going to fix this without changing the text of the standard.
So I just try to replace std::prev with other functional equivalents.