Fix config2.xsd
I got an error:
```
$ xmllint --noout --schema config/config2.xsd test/config/fixtures/mock-config-all.xml
config/config2.xsd:1183: element choice: Schemas parser error :
 Element '{http://www.w3.org/2001/XMLSchema}all': The content is not valid.
 Expected is (annotation?, (annotation?, element*).
WXS schema config/config2.xsd failed to compile
```

Two different mimetype variants shall be supported:
```
<profile name="a">
    <mimetype>image/png</mimetype>
</profile>

<profile name="b">
    <mimetype value="image/png">
        <mime-property key="foo" resource="bar" metadata="baz"/>
    </mimetype>
</profile>
```

but we are limited in validating this with XML Schema 1.0:
https://stackoverflow.com/questions/381782/xml-schema-element-that-can-contain-elements-or-text

So mimetype is declared to be a mixed content tag,
allowing text values as well as mime-property subtags.

Unfortunately this also allows
<mimetype></mimetype>
<mimetype><mime-property key="foo"/></mimetype>
which both misses information we need, but that could only
be solved with XML Schema 1.1, which is not supported by xmllint.